convert Divider to ts
