--- src/Divider/index.js
@@ -1,2 +1,0 @@
-import Divider from './Divider';
-export default Divider;
