+++ src/Divider/index.ts
@@ -1,0 +1,3 @@
+import Divider, { DividerProps } from './Divider';
+export default Divider;
+export { DividerProps };
